Move migration functions to a separate module
... and split it up into smaller functions
