+++ crates/matrix-sdk-indexeddb/src/crypto_store/mod.rs
@@ -43,7 +43,9 @@ use tokio::sync::Mutex;
 use wasm_bindgen::JsValue;
 use web_sys::IdbKeyRange;
 
-use crate::safe_encode::SafeEncode;
+use crate::{crypto_store::migrations::open_and_upgrade_db, safe_encode::SafeEncode};
+
+mod migrations;
 
 mod keys {
     // stores
@@ -155,97 +157,7 @@ impl IndexeddbCryptoStore {
     ) -> Result<Self> {
         let name = format!("{prefix:0}::matrix-sdk-crypto");
 
-        // Open my_db v1
-        let mut db_req: OpenDbRequest = IdbDatabase::open_u32(&name, 5)?;
-        db_req.set_on_upgrade_needed(Some(|evt: &IdbVersionChangeEvent| -> Result<(), JsValue> {
-            // Even if the web-sys bindings expose the version as a f64, the IndexedDB API
-            // works with an unsigned integer.
-            // See <https://github.com/rustwasm/wasm-bindgen/issues/1149>
-            let old_version = evt.old_version() as u32;
-
-            if old_version < 1 {
-                // migrating to version 1
-                let db = evt.db();
-
-                db.create_object_store(keys::CORE)?;
-                db.create_object_store(keys::SESSION)?;
-
-                db.create_object_store(keys::INBOUND_GROUP_SESSIONS)?;
-                db.create_object_store(keys::OUTBOUND_GROUP_SESSIONS)?;
-                db.create_object_store(keys::TRACKED_USERS)?;
-                db.create_object_store(keys::OLM_HASHES)?;
-                db.create_object_store(keys::DEVICES)?;
-
-                db.create_object_store(keys::IDENTITIES)?;
-                db.create_object_store(keys::BACKUP_KEYS)?;
-            }
-
-            if old_version < 2 {
-                let db = evt.db();
-
-                // We changed how we store inbound group sessions, the key used to
-                // be a trippled of `(room_id, sender_key, session_id)` now it's a
-                // tuple of `(room_id, session_id)`
-                //
-                // Let's just drop the whole object store.
-                db.delete_object_store(keys::INBOUND_GROUP_SESSIONS)?;
-                db.create_object_store(keys::INBOUND_GROUP_SESSIONS)?;
-                db.create_object_store(keys::ROOM_SETTINGS)?;
-            }
-
-            if old_version < 3 {
-                let db = evt.db();
-
-                // We changed the way we store outbound session.
-                // ShareInfo changed from a struct to an enum with struct variant.
-                // Let's just discard the existing outbounds
-                db.delete_object_store(keys::OUTBOUND_GROUP_SESSIONS)?;
-                db.create_object_store(keys::OUTBOUND_GROUP_SESSIONS)?;
-
-                // Support for MSC2399 withheld codes
-                db.create_object_store(keys::DIRECT_WITHHELD_INFO)?;
-            }
-
-            if old_version < 4 {
-                let db = evt.db();
-
-                db.create_object_store(keys::SECRETS_INBOX)?;
-            }
-
-            if old_version < 5 {
-                let db = evt.db();
-
-                // Create a new store for outgoing secret requests
-                let object_store = db.create_object_store(keys::GOSSIP_REQUESTS)?;
-
-                let mut params = IdbIndexParameters::new();
-                params.unique(false);
-                object_store.create_index_with_params(
-                    keys::GOSSIP_REQUESTS_UNSENT_INDEX,
-                    &IdbKeyPath::str("unsent"),
-                    &params,
-                )?;
-
-                let mut params = IdbIndexParameters::new();
-                params.unique(true);
-                object_store.create_index_with_params(
-                    keys::GOSSIP_REQUESTS_BY_INFO_INDEX,
-                    &IdbKeyPath::str("info"),
-                    &params,
-                )?;
-
-                if old_version > 0 {
-                    // we just delete any existing requests.
-                    db.delete_object_store("outgoing_secret_requests")?;
-                    db.delete_object_store("unsent_secret_requests")?;
-                    db.delete_object_store("secret_requests_by_info")?;
-                }
-            }
-
-            Ok(())
-        }));
-
-        let db: IdbDatabase = db_req.await?;
+        let db = open_and_upgrade_db(&name).await?;
         let session_cache = SessionStore::new();
 
         Ok(Self {
